+++ packages/shared/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./paths.js";
 export * from "./config-loader.js";
 export * from "./logger.js";
+export * from "./protocol-sync.js";